fix issue #1 (Does not seem to understand the Kasa HS300 power strips)
The log file clearly showed that the message received from the device
was malformed json, and therefore crashed the parser.  It was
malformed because it was incompletely received.  The truncation was
due to a maxlen parameter that was simply too small.
The fix is to increase this value (now arbitrarily set to 4096).

In addition, I factored two magic numbers into constants to make the
code more understandable.